dispatcher/requirements.txt
# Airflow核心依赖 (升级到最新稳定版本)
apache-airflow==2.11.0

# Maxcompute Provider (支持Maxcompute Connections)
apache-airflow-providers-alibaba[cloud]>=3.2.0,<4.0.0

# Maxcompute SDK
pyodps>=0.11.0

# 数据库连接
pymysql>=1.0.2
sqlalchemy>=1.4.0,<2.0.0

# 数据处理
pandas>=1.5.0
numpy>=1.21.0

# 工具库
python-dotenv>=0.19.0
requests>=2.28.0

# 测试相关
pytest>=7.0.0
pytest-cov>=4.0.0

# 开发工具
black>=22.0.0
flake8>=5.0.0
mypy>=1.0.0 